% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{actual_vp}
\alias{actual_vp}
\title{Actual Vapor Pressure}
\usage{
actual_vp(rhmax, rhmin, tmax, tmin)
}
\arguments{
\item{rhmax}{Daily maximum relative humidity (percent).}

\item{rhmin}{Daily minimum relative humidity (percent).}

\item{tmax}{Daily maximum temperature (deg C).}

\item{tmin}{Daily minimum temperature (deg C).}
}
\description{
Calculates actual vapor pressure (kPa) of air based on maximum and minimum relative humidity and maximum and minimum temperature.
}
