% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{atm_press}
\alias{atm_press}
\title{Atmospheric Pressure}
\usage{
atm_press(elev)
}
\arguments{
\item{elev}{Elevation (m).}
}
\description{
Estimates atmospheric pressure (kPa) at a given elevation.
}
