% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{clear_sky_rad}
\alias{clear_sky_rad}
\title{Clear Sky Radiation}
\usage{
clear_sky_rad(doy, lat, elev)
}
\arguments{
\item{doy}{Day-of-year (Julian date)}

\item{lat}{Latitude (degrees)}

\item{elev}{Elevation (m)}
}
\description{
Calculates incoming clear-sky radiation (MJ m^-2 day^-1) based on day-of-year, latitude, and elevation
}
