% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{ET_Hamon_daily}
\alias{ET_Hamon_daily}
\title{Hamon Daily PET}
\usage{
ET_Hamon_daily(x)
}
\arguments{
\item{x}{A daily time series data frame containing tmean_C (deg C), and daylength (hours)}
}
\description{
Calculates Hamon PET from a daily time series of Tmean and daylength.
}
