% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{ET_PenmanMonteith_daily}
\alias{ET_PenmanMonteith_daily}
\title{Penman-Monteith Daily PET}
\usage{
ET_PenmanMonteith_daily(x, elev, lat, wind = NULL)
}
\arguments{
\item{x}{A daily time series data frame containing Date (date object), tmax_C (deg C), tmin_C (deg C), srad (MJ m^-2 day^-1). Optionally contains RHmax (percent), RHmin (percent), vp (kPa), and wind (m/s).}

\item{elev}{Elevation of the site (m).}

\item{lat}{Latitude of the site (degrees).}

\item{wind}{(optional) An estimated value for daily average wind speeds (m/s). Use if input data frame does not contain daily wind speed values.}
}
\description{
Calculates PET (mm) from daily Tmax, Tmin, solar radiation, elevation, and latitude, according to the Penman-Monteith method. May also use daily maximum and minimum relative humidity, atmospheric vapor pressure, and wind speeds.
}
