% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{ET_Thorn_monthly}
\alias{ET_Thorn_monthly}
\title{Thornthwaite Monthly PET}
\usage{
ET_Thorn_monthly(x)
}
\arguments{
\item{x}{A monthly time series data frame containing Date, tmean_C (deg C), and daylength (hours)}
}
\description{
Calculates PET from monthly Tmean and daylength, according to the Thornthwaite method.
}
