% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{fold_aspect}
\alias{fold_aspect}
\title{Folded aspect}
\usage{
fold_aspect(aspect)
}
\arguments{
\item{aspect}{Aspect in degrees (0-360)}
}
\description{
"Folds" aspect over N-S line, so aspects are limited to 0-180
degrees
}
