% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_AET}
\alias{get_AET}
\title{Actual Evapotranspiration (AET)}
\usage{
get_AET(w, pet, swc, swc.0)
}
\arguments{
\item{w}{A time series vector of available water for soil charging (rain + snowmelt).}

\item{pet}{A time series vector of PET.}

\item{swc}{A time series vector of soil water content.}

\item{swc.0}{(optional) The initial soil water content value. Default is 0.}
}
\description{
Calculates actual evapotranspiration (AET) from available water, PET, and soil water.
}
