% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_d_soil}
\alias{get_d_soil}
\title{Daily change in Soil Water Content (SWC)}
\usage{
get_d_soil(swc, swc.0 = NULL)
}
\arguments{
\item{swc}{A time series vector of soil water content.}

\item{swc.0}{(optional) The initial soil water content value. Default is 0.}
}
\description{
Calculates daily change in soil water content.
}
