% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{get_daylength}
\alias{get_daylength}
\title{Daylength}
\usage{
get_daylength(dates, lat)
}
\arguments{
\item{dates}{A series of dates containing year, month, and day}

\item{lat}{Latitude (degrees)}
}
\description{
Returns daylength in hours for a series of dates, based on latitude. Calls the 'geosphere' package.
}
