% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_deficit}
\alias{get_deficit}
\title{Climatic water deficit}
\usage{
get_deficit(pet, AET)
}
\arguments{
\item{pet}{A time series vector of PET.}

\item{AET}{A time series vector of actual evapotranspiration.}
}
\description{
Calculates daily climatic water deficit, which is PET - AET.
}
