% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_freeze}
\alias{get_freeze}
\title{Freeze factor using Jennings et al., 2018 thresholds to partition rain and snow}
\usage{
get_freeze(jtemp, tmean)
}
\arguments{
\item{jtemp}{the Jennings temperature extracted from the raster based on latitude and longitude.}

\item{tmean}{A vector of daily mean temperatures (deg C).}
}
\description{
Calculates a freeze factor from 0-1 based on a temperature threshold from Jennings et al., 2018 and average temperature.
}
