% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_GDD}
\alias{get_GDD}
\title{Growing Degree-Days}
\usage{
get_GDD(tmean, tbase)
}
\arguments{
\item{tmean}{A time series vector of daily mean temperatures (deg C)}

\item{tbase}{(optional) A threshold temperature, above which growing degree-days are calculated. Default is 0.}
}
\description{
Calculates growing degree-days at daily time steps based on mean temperature and a threshold temperature
}
