% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{get_heatload}
\alias{get_heatload}
\title{Heatload}
\usage{
get_heatload(slope, aspect, latitude)
}
\arguments{
\item{slope}{Slope in degrees, from 1-60 degrees.}

\item{aspect}{Aspect in degrees.}

\item{latitude}{Latitude in degrees}
}
\description{
Estimate potential annual direct incident radiation from slope,
aspect, and latitude.  From McCune, B., and Keon, D., 2002,
Equations for Potential Annual Direct Incident Radiation and Heat
Load: Journal of Vegetation Science, v. 13, p. 603–606.
}
