% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_jtemp}
\alias{get_jtemp}
\title{Temperature threshold using Jennings et al. 2018 to partition rain and snow}
\usage{
get_jtemp(Lon, Lat)
}
\arguments{
\item{Lon}{Longitude of the site (degrees).}

\item{Lat}{Latitude of the site (degrees).}
}
\description{
Extracts the rain-snow temperature threshold from a raster.
}
