% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_melt}
\alias{get_melt}
\title{Melt}
\usage{
get_melt(tmean, j_temp, hock, snow, sp.0 = NULL)
}
\arguments{
\item{tmean}{A vector of daily mean temperatures (deg C).}

\item{j_temp}{the Jennings temperature extracted from the raster based on latitude and longitude}

\item{hock}{A melt factor of daily snowmelt when warm enough to melt.}

\item{snow}{A time series vector of snowfall values.}

\item{snowpack}{A time series vector of snowpack accumulation values. Initial snowpack default value is 0.}
}
\description{
Calculates the amount of snowmelt at time steps from snowpack, temperature, and Hock melt factor
}
