% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{get_OudinPET}
\alias{get_OudinPET}
\title{Oudin Daily PET}
\usage{
get_OudinPET(doy, lat, snowpack, tmean, slope, aspect, shade.coeff = NULL)
}
\arguments{
\item{doy}{Day-of-year (Julian date)}

\item{lat}{Latitude of the site (degrees).}

\item{snowpack}{A time series vector of snowpack accumulation values.}

\item{tmean}{A vector of daily mean temperatures (deg C).}

\item{slope}{Slope of the site (in degrees).}

\item{aspect}{Aspect of the site (in degrees).}

\item{shade.coeff}{(optional) A shade coefficient from 0-1. Default is 1.}
}
\description{
Calculates PET (mm) based on temperature, latitude, and solar radiation
}
