% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_rain}
\alias{get_rain}
\title{Rain}
\usage{
get_rain(ppt, freeze)
}
\arguments{
\item{ppt}{A vector of precipitation values.}

\item{freeze}{A vector of freeze factor values, calculated from average temperature and Jennings et al., 2018. Values are 0-1.}
}
\description{
Calculates rainfall totals based on precipitation and freeze factor.
}
