% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{get_rh}
\alias{get_rh}
\title{Relative Humidity}
\usage{
get_rh(vp, temp)
}
\arguments{
\item{vp}{Vapor pressure (kPa)}

\item{temp}{Temperature (deg C)}
}
\description{
Calculates relative humidity (%) from atmospheric vapor pressure and temperature
}
