% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_runoff}
\alias{get_runoff}
\title{Runoff or excess input greater than soil water holding capacity}
\usage{
get_runoff(ppt, w, d_soil, AET, R.coeff = NULL)
}
\arguments{
\item{ppt}{A vector of precipitation values.}

\item{w}{A time series vector of available water for soil charging (rain + snowmelt).}

\item{d_soil}{A time series vector of change in soil moisture from previous day.}

\item{AET}{A time series vector of actual evapotranspiration.}

\item{R.coeff}{A fraction of precpitation that can be shunted to direct runoff.}
}
\description{
Calculates runoff at daily timesteps based on water reaching soil surface, AET, change in soil moisture, and a runoff coefficient
}
