% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_snow}
\alias{get_snow}
\title{Snow}
\usage{
get_snow(ppt, freeze)
}
\arguments{
\item{ppt}{A vector of precipitation values.}

\item{freeze}{A vector of freeze factor values, calculated from Tmean. Values are 0-1.}
}
\description{
Calculates snowfall totals based on precipitation and freeze factor
}
