% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{get_svp}
\alias{get_svp}
\title{Saturation Vapor Pressure}
\usage{
get_svp(temp)
}
\arguments{
\item{temp}{Temperature (deg C)}
}
\description{
Calculates mean saturation vapor pressure (kPa) of air based on temperature (deg C).
}
