% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_w}
\alias{get_w}
\title{Water reaching soil surface as snow plus rain}
\usage{
get_w(rain, melt)
}
\arguments{
\item{rain}{A vector of daily rain.}

\item{melt}{A vector of daily snowmelt.}
}
\description{
Calculates water reaching soil surface using rain and snowmelt.
}
