% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_w_pet}
\alias{get_w_pet}
\title{Water reaching soil surface minus PET}
\usage{
get_w_pet(w, pet)
}
\arguments{
\item{w}{A time series vector of water reaching soil surface as snow plus rain.}

\item{pet}{A time series vector of PET values.}
}
\description{
Calculates water reaching soil surface minues the PET.
}
