% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{outgoing_rad}
\alias{outgoing_rad}
\title{Outgoing Radiation}
\usage{
outgoing_rad(tmax, tmin, R.s, e.a, R.so)
}
\arguments{
\item{tmax}{Daily maximum temperatures (deg C).}

\item{tmin}{Daily minimum temperatures (deg C).}

\item{R.s}{Incoming solar radiation (MJ m^-2 day^-1).}

\item{e.a}{Actual vapor pressure (kPa).}

\item{R.so}{Clear-sky radiation (MJ m^-2 day^-1).}
}
\description{
Calculates outgoing radiation (MJ m^-2 day^-1) based on daily Tmax, Tmin, incoming radiation, actual vapor pressure, and clear-sky radiation.
}
