% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{psyc_constant}
\alias{psyc_constant}
\title{Psychrometric Constant}
\usage{
psyc_constant(elev)
}
\arguments{
\item{elev}{Elevation (m)}
}
\description{
Calculates the psychrometric constant relating partial pressure of water in air to the air temperature, based on atmospheric pressure. Calls the atm_press() function to estimate atmospheric pressure from elevation.
}
