% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{sat_vp}
\alias{sat_vp}
\title{Saturation Vapor Pressure}
\usage{
sat_vp(temp)
}
\arguments{
\item{temp}{A vector or single value of temperatures (deg C)
sat_vp()}
}
\description{
Calculates mean saturation vapor pressure of air based on temperature.
}
