% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_functions.R
\name{vapor_curve}
\alias{vapor_curve}
\title{Slope of Saturation Vapor Curve}
\usage{
vapor_curve(temp)
}
\arguments{
\item{temp}{A time series vector or single value of temperatures (deg C).}
}
\description{
Calculates the slope of the saturation vapor curve for a given temperature.
}
