% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pet.R
\name{wb_pet}
\alias{wb_pet}
\title{Calculate Potential Evapo-Transpiration}
\usage{
wb_pet(
  solar,
  temp,
  month,
  humid = T,
  rel_hum = NULL,
  pet_adjst = T,
  pet_adjst_coefs = NULL
)
}
\arguments{
\item{solar}{Gridded total monthly radiation in Wh/m2 units.}

\item{temp}{Gridded average monthly temperature in degrees Celsius units}

\item{month}{The abbreviated name of the month associated with these raster
data. Valid abbreviations must be three letters, and match those in
\code{month.abb}}

\item{humid}{Logical variable indicating whether the study region is in a
humid climate or not. Defaults to true, since the Turc method is designed
to be run for humid climates. If set to FALSE, the \code{rel_hum} argument must
be provided.}

\item{rel_hum}{Gridded average monthly relative humidity layer, used to apply
the correction factor for calculating PET in non-humid climates.}

\item{pet_adjst}{T/F indicating whether a PET adjustment factor should be
applied to the calculated PET. If set to TRUE then the \code{pet_adjst_coefs}
argument must be provided.}

\item{pet_adjst_coefs}{Gridded adjustment coefficients to be applied to the
calculated PET layer.}
}
\value{
A \code{terra} raster layer giving the calculated PET.
}
\description{
Given a gridded radiation layer and a monthly average temperature layer, this
function calcualtes the Potential Evapotranspiration for each grid cell in
the study region. Input rasters must match in resolution, extent and CRS.
}
\details{
The function is built on the \code{terra} geospatial library. Other raster format
such as \code{raster} and \code{stars} will also be acceptedu, but will be converted to
\code{terra} under the hood for computation. It is recommended you convert to
\code{terra} before calling the function to minimize side effects from
unsupervised conversion.
}
\examples{
# Working on it
}
