% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_precip_minus_pet.R
\name{wb_precip_minus_pet}
\alias{wb_precip_minus_pet}
\title{Calculate Precipitation Minus Potential Evapo-Transpiration}
\usage{
wb_precip_minus_pet(precip, pet)
}
\arguments{
\item{precip}{Gridded total monthly radiation in Wh/m2 units.}

\item{pet}{Gridded average monthly temperature in degrees Celsius units}
}
\value{
A \code{terra} raster layer giving the calculated P-PET.
}
\description{
Given a gridded precipitation layer and a PET layer, this difference between
precipitation and PET
}
\examples{
# Working on it
}
