% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_soil_storage.R
\name{wb_soil_storage_annual}
\alias{wb_soil_storage_annual}
\title{Calculate Soil Storage for a Full Year of Data}
\usage{
wb_soil_storage_annual(
  awc,
  precip_stack,
  p_pet_stack,
  prev_storage = NULL,
  wrap = T,
  fileOrderWarning = T
)
}
\arguments{
\item{awc}{Gridded average soil water holding capacity.}

\item{precip_stack}{List object containing 12 raster layers giving monthly
total precipitation (mm) for each month in the year. The list must be
ordered from January to December.}

\item{p_pet_stack}{List object containing 12 raster layers giving the
difference between monthly total precipitation and potential
evapo-transpiration (mm) for each month in the year. The list must be
ordered from January to December.}

\item{prev_storage}{A raster layer defining soil storage in the month prior
to the starting month. See Details.}

\item{wrap}{Boolean indicating whether the soil storage calculation should
"wrap" around after calculating storage for the last month. Defaults to
TRUE. See Details.}

\item{fileOrderWarning}{Should the user be warned about input layers stacks
being ordered from January to December. Defaults to TRUE.}
}
\value{
A \code{terra} raster layer giving the calculated P-PET.
}
\description{
Given a gridded precipitation layer and a PET layer, this difference between
precipitation and PET
}
\examples{
# Working on it
}
