% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_soil_storage.R
\name{wb_soil_storage_month}
\alias{wb_soil_storage_month}
\title{Calculate Soil Storage For a Specific Month}
\usage{
wb_soil_storage_month(awc, precip, p_pet, prev_storage)
}
\arguments{
\item{awc}{Gridded average soil water holding capacity.}

\item{precip}{Gridded total monthly radiation in Wh/m2 units.}

\item{p_pet}{Gridded difference between precipitation and potential
evapo-transporation (PET)}

\item{prev_storage}{A raster layer defining soil storage in the month
preceding this month. See Details.}
}
\value{
A \code{terra} raster layer giving the calculated P-PET.
}
\description{
Given a gridded precipitation layer and a PET layer, this difference between
precipitation and PET
}
\examples{
# Working on it
}
