% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_freeze}
\alias{get_freeze}
\title{Freeze factor}
\usage{
get_freeze(tmean)
}
\arguments{
\item{tmean}{A vector of daily mean temperatures (deg C).}
}
\description{
Calculates a freeze factor from 0-1 based on mean temperature
}
