% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_melt}
\alias{get_melt}
\title{Melt}
\usage{
get_melt(tmean, jtemp, hock, snow, sp.0 = NULL)
}
\arguments{
\item{tmean}{A vector of daily mean temperatures (deg C).}

\item{jtemp}{the Jennings temperature extracted from the raster based on latitude and longitude.}

\item{hock}{A melt factor of daily snowmelt when warm enough to melt.}

\item{snow}{A time series vector of snowfall values.}

\item{sp.0}{(optional) Initial snowpack value. Default is 0.}
}
\description{
Calculates the amount of snowmelt at time steps from snowpack, temperature, and Hock melt factor.
}
