% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_melt}
\alias{get_melt}
\title{Melt}
\usage{
get_melt(snowpack, snow, freeze, p.0 = NULL)
}
\arguments{
\item{snowpack}{A time series vector of snowpack accumulation values.}

\item{snow}{A time series vector of snowfall values.}

\item{freeze}{A vector of freeze factor values, calculated from Tmean. Values are 0-1.}

\item{p.0}{(optional) Initial snowpack value. Default is 0.}
}
\description{
Calculates the amount of snowmelt at time steps from snowpack, snowfall, and freeze factor.
}
