% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_snowpack}
\alias{get_snowpack}
\title{Snowpack}
\usage{
get_snowpack(ppt, freeze, p.0 = NULL)
}
\arguments{
\item{ppt}{A vector of precipitation values.}

\item{freeze}{A vector of freeze factor values, calculated from Tmean. Values are 0-1.}

\item{p.0}{(optional) Initial snowpack value. Default is 0.}
}
\description{
Calculates snowpack accumulation at time steps, from a time series of precipitation values, freeze factor, and an initial snowpack value
}
