% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{get_snowpack}
\alias{get_snowpack}
\title{Snowpack}
\usage{
get_snowpack(snow, melt, sp.0 = NULL)
}
\arguments{
\item{snow}{A time series vector of snowfall values.}

\item{melt}{A time series vector of snowmelt.}

\item{sp.0}{(optional) Initial snowpack value. Default is 0.}
}
\description{
Calculates snowpack accumulation at time steps, from a time series of snowfall and melt.
}
